via muerta, parece que bash no expande antes del análisis sintactico, pero si antes de dividir los argumentos para exec argv
