+++ APUNTES/NOTAS
@@ -36,3 +36,5 @@ valgrind --leak-check=full --show-leak-kinds=all --suppressions=readline.supp ./
 
 15feb TEST ROTOS HOY
 Test 132
+
+$CONESPACIOINICIAL$CONESPACIOINICIAL es cos
