+++ paluego_jose/expansor_utils.c
