+++ parser/1_expansor/expansor.c
@@ -1,102 +1,44 @@
-
 #include "../../minishell.h"
 #include "../../libft/headers/libft.h"
 
-
-
-
-static int	prepare_expansion(char **origin, char **new_str, char *envp[])
+//consume marker hasta el siguiente '$', y retorna una copia de la cadena consumida.
+// Si marker apuntaba a '$' consumira los caracteres a continuacion hasta encontrar
+// un caracter separador (<,>,|, ,) y retornará una copia de la cadena, si la variable existe.
+// si la variable no existe retorna una cadena vacia.
+// si $ es seguido de un separador (por ejemplo "$|") consume $ y retorna la cadena "$\0"
+char *get_piece(char **marker, char *envp[])
 {
-	int	len;
+	char *piece;
 
-	len = calculate_expansion_length(*origin, envp);
-	if (len < 0)
-		return (1);
-	*new_str = ft_calloc(len + 1, sizeof(char));
-	if (*new_str == NULL)
-	 	return (1);
-//
-	*origin = *new_str;
-	(*new_str)[len] = '\0';
-	return (0);
+	piece = ft_substr(*marker, 0, );
+
 }
 
-
-e_errors	expandstr(char **origin, char *envp[]) //envp debe recibir el array de strings que hemos creado y sobre el que se reflejan las modificaciones que pueda hacer minishell durante la ejecucion
+e_errors	expandstr(char **origin, char *envp[])
 {
-	char	*marker;
-	char	*new_str;
-	char	*str;
+	char *newstr;
+	char *auxstr;
+	char *piece;
+	char *marker;
 
-	if(!is_expansible(*origin))
-		return (ALL_OK);
-	str = *origin;
-	if (prepare_expansion(origin, &new_str, envp))
+	marker = *origin;
+	newstr = get_piece(&marker, envp);
+	if(!newstr)
 		return (ERROR_MALLOC);
-	marker = str;
-	while (*marker)
+	while(strchr(marker, '$'))
 	{
-		if (*marker == '$')
-		{
-			if(handle_dollar(&new_str, &str, &marker, envp))
-				return (ERROR_MALLOC);
-			continue ;
-		}
-		if (*marker == 39 && ft_strchr(marker + 1, 39))
-			marker = ft_strchr(marker + 1, 39);
-		marker++;
+		piece = get_piece(&marker, envp);
+		if(!piece)
+			return (ERROR_MALLOC);
+		//anadir fragmento a la cadena
+		auxstr = newstr;
+		newstr = ft_strjoin(newstr, piece);
+		free(piece);
+		free(auxstr);
+		if(!newstr)
+			return (ERROR_MALLOC);
 	}
-	if (*str)
-        ft_strlcpy(new_str, str, marker - str + 1);
-    return (ALL_OK);
-}
-
-//
-/*
-static int	expand_task(t_task *node, char *envp[])
-{
-	int	i;
-
-	node->garb.size = count_expansions(((t_task *)node)) + 1;
-	node->garb.pointers = ft_calloc(node->garb.size, sizeof(void *));
-	if (node->garb.pointers == NULL)
-		return (1);
-	(node->garb.pointers)[node->garb.size -1] = NULL;
-	
-	if (expandstr(&(node->cmd), envp))
-		return (1);
-//	if (add_pathname(&(node->cmd), &(node->garb), envp))
-//		return (1);
-	if (expandstr(&(node->redir.infoo), envp))
-		return (1);
-	if (expandstr(&(node->redir.outfile), envp))
-		return (1);
-	i = 0;
-	while ((node->argv)[i])
-	{
-		if (expandstr((&(node->argv)[i]), envp))
-			return (1);
-		i++;
-	}
-	return(0);
-}
-*/
-
-//verifica si es necesaria la expansion, si lo es, la realiza llamando a expandstr y libera la linea original
-e_errors expansor(char **line, char **envp)
-{
-	char	*oldline;
-
-	oldline = *line;
-	if (is_expansible(*line))
-	{
-		oldline = *line;
-		if (expandstr(line, envp))
-		{
-			free(oldline);
-			return(ERROR_MALLOC);
-		}
-		free(oldline);
-	}
-	return(ALL_OK);
+	free(origin);
+	*origin = newstr;
+	return (ALL_OK);
 }
